test/helm: split cross-chart helpers into shared_test.go
